+++ game/Lmd_Prof_Merc.h
@@ -19,3 +19,6 @@ int Lmd_Prof_Merc_GetLethalitySkill(Account_t *acc);
 
 int Lmd_Prof_Merc_Getrifle_masterSkill(Account_t *acc);
 #define PlayerProf_Merc_Getrifle_masterSkill(ent) Lmd_Prof_Merc_Getrifle_masterSkill(ent->client->pers.Lmd.account)
+
+int Lmd_Prof_Merc_Getperfect_aimSkill(Account_t *acc);
+#define PlayerProf_Merc_Getperfect_aimSkill(ent) Lmd_Prof_Merc_Getperfect_aimSkill(ent->client->pers.Lmd.account)